fix(format): use portable format specifiers for int64_t/uint64_t
On LP64 (64-bit Linux/macOS), int64_t is long (8 bytes), not long long.
Using %lld for int64_t arguments causes -Wformat warnings because the
format specifier expects unsigned long long but the argument type is long.

Fix by casting int64_t arguments to (long long) and uint64_t to
(unsigned long long) at format string call sites. Also fix:
- PDSCLIDispatcher: cast data.bytes to (const char *) for %s
- WebSocketServer: use gai_strerror(gai_err) instead of raw int for %s

+++ Garazyk/Sources/Database/Service/ServiceDatabases.m
@@ -828,7 +828,7 @@ static NSString *refreshTokenSessionID(NSString *refreshToken) {
     // Defense-in-depth: warn if seq is invalid (should be positive per ATProto spec)
     if (seq <= 0) {
         PDS_LOG_SYNC_WARN(@"persistEvent called with invalid seq=%lld for type=%@; "
-                           @"this indicates a sequence number bug", seq, type);
+                           @"this indicates a sequence number bug", (long long)seq, type);
     }
     __block BOOL success = NO;
     [self.sequencerPool transactWithDid:@"__service__" block:^(id<PDSActorStoreTransactor> transactor, NSError **innerError) {
